% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_raven.R
\name{exp_raven}
\alias{exp_raven}
\title{Export raven selections}
\usage{
exp_raven(X, path = NULL, file.name = NULL, khz.to.hz = TRUE, 
sound.file.path = NULL, single.file = TRUE)
}
\arguments{
\item{X}{Data frame containing columns for sound file (sound.files), selection (selec), start and end time of signals ('start' and 'end') and low and high frequency ('low.freq' and 'high.freq', optional). See example data 'selec.table' in the \code{\link{warbleR}}) package.}

\item{path}{A character string indicating the path of the directory in which to save the selection files. 
If not provided (default) the function saves the file into the current working directory.}

\item{file.name}{Name of the output .txt file. If \code{NULL} then the sound file names are used instead. If multiple
selection files are generated (see 'single.file') then the sound files names are added to the provided 'file.name'.}

\item{khz.to.hz}{Logical. Controls if frequency variables should be converted from kHz (the unit used by other bioacoustic analysis R packages like \code{\link{warbleR}}) to Hz (the unit used by Raven). Default is \code{TRUE}.}

\item{sound.file.path}{A character string indicating the path of the 
directory containing the sound file(s). Providing this information allows
 to open both sound file and selection table simultaneously. This can be
 done by using the 'File > Open selection table' option in Raven (or drag/drop the 
selection file into Raven). Default is \code{NULL}. This argument is required when
exporting selections from multiple sound files.}

\item{single.file}{Logical. Controls whether a single selection file (\code{TRUE}; default)
or multiple selection files for each sound files (\code{FALSE}, hence, only applicable
when several sound files are included in 'X') are generated. Note that
'sound.file.path' must be provided when exporting several sound files into a single selection file as the
duration of the sound files is required.}
}
\value{
The function saves a selection table in '.txt' format that can be 
directly opened in Raven. If several sound files are available users can either 
export them as a single selection file or as multiple selection files (one for each sound file). 
No objects are returned in the R environment.
}
\description{
\code{exp_raven} exports selection tables as Raven selection data in .txt format.
}
\details{
The function exports selection tables as the ones used by the bioacoustic analysis R package \code{\link{warbleR}} to Raven selection files in '.txt' format. This can be useful to obtain additional Raven
measurements on existing selections by adding new measurements to the 
selection table once in Raven. Note that selection labels must be numeric and unduplicated 
when exporting them to Raven. If that is not the case the function will
relabeled the selections and the previous selection labels will be retained in a new column('old.selec').
}
\examples{
\dontrun{
# First set temporary folder
setwd(tempdir())

# Load data
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))

# Select data for a single sound file
st1 <- selec.table[selec.table$sound.files == "Phae.long1.wav",]

# Export data of a single sound file
exp_raven(st1, file.name = "Phaethornis Rraven examples")

writeWave(Phae.long1, "Phae.long1.wav") #save sound files 
writeWave(Phae.long2, "Phae.long2.wav")
writeWave(Phae.long3, "Phae.long3.wav")
writeWave(Phae.long4, "Phae.long4.wav")

exp_raven(X = selec.table, file.name = "Phaethornis warbleR examples", 
sound.file.path = tempdir(), single.file = T)
}

}
\author{
Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}
\seealso{
\code{\link{imp_raven}}; \code{\link{imp_syrinx}}
}

