% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rraven-package.R
\docType{package}
\name{Rraven}
\alias{Rraven}
\alias{Rraven-package}
\title{warbleR: Exchange data and open sound files in Raven from R}
\description{
Rraven is a package designed to faciliate the exchange of data between R 
and the Raven bioacoustics analysis software.
}
\details{
Functions:

License: GPL (>= 2)
}
\section{Obtaining avian vocalization data}{

  
  \code{\link{imp.raven}}: Importing Raven selections
  
  \code{\link{exp_raven}}: Export R selection tables into Raven selection file format

  \code{\link{run_raven}}: Open sound files in Raven

  \code{\link{imp_syrinx}}: Importing Syrinx selections
}
\author{
Marcelo Araya-Salas
  
  Maintainer: Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}

