% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_file_ts-data.R
\docType{data}
\name{selection_file_ts}
\alias{selection_file_ts}
\title{Data frame of selections imported from Raven including a time series parameter.}
\format{A data frame with 3 rows and 14 variables}
\source{
Marcelo Araya-Salas
}
\usage{
data(selection_file_ts)
}
\description{
A data frame containing the acoustic parameters measured on 
\emph{Phaethornis longirostris} (Long-billed Hermit) songs from the 
example sound files included in the \code{\link{warbleR}} package.
}
\keyword{datasets}

