% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_files.R
\docType{data}
\name{selection_files}
\alias{selection_files}
\title{A list of Raven selection tables.}
\format{List of 5 data frame. The last data frame contains a multiple sound file selection table.}
\source{
Marcelo Araya-Salas
}
\usage{
data(selection_files)
}
\description{
List of Raven selection tables containing 
\emph{Phaethornis longirostris} (Long-billed Hermit) songs from the 
example sound files included in the \code{\link{warbleR}} package.
}
\keyword{datasets}

