% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rraven-package.R
\docType{package}
\name{Rraven}
\alias{Rraven}
\alias{Rraven-package}
\title{Rraven: Exchange data and open sound files in 'Raven' from R}
\description{
Rraven is a package designed to facilitate the exchange of data
between R and 'Raven' sound analysis software (Cornell Lab of Ornithology)
}
\details{
License: GPL (>= 2)  

 @section Functions:
  
  \code{\link{extract_ts}}: Extract time series parameters from data imported from Raven
  
  \code{\link{exp_raven}}: Export R selection tables into 'Raven' selection file format
  
  \code{\link{imp_corr_mat}}: Import 'Raven' batch correlator output 
  
  \code{\link{imp_raven}}: Importing 'Raven' selections
  
  \code{\link{imp_syrinx}}: Importing 'Syrinx' selections
  
  \code{\link{relabel_colms}}: Relabel columns to match the selection table format
  
  \code{\link{run_raven}}: Open sound files in Raven
}
\author{
Marcelo Araya-Salas
  
  Maintainer: Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}
