% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_files.R
\docType{data}
\name{selection_files}
\alias{selection_files}
\title{A list of Raven selection tables.}
\format{List of 6 selection tables. The fifth table contains a 
multiple sound file selection table. The last table is formated as 
in Syrinx.}
\source{
Marcelo Araya-Salas
}
\usage{
data(selection_files)
}
\description{
List of Raven and Syrinx selection tables containing 
\emph{Phaethornis longirostris} (Long-billed Hermit) songs from the 
example sound files included in the \code{\link{warbleR}} package 
and one selection from cane-brake wrens (no sound file available).
}
\keyword{datasets}

