% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_files.R
\docType{data}
\name{selection_files}
\alias{selection_files}
\title{A list of 'Raven' selection tables.}
\format{
List of 7 selection tables. The first 5 items are 'Raven' selection tables of Long billed hermit songs. The 5th table contains a multiple sound file selection 
table. The 6th item is a 'Raven' correlation matrix and the last one a table in 'Syrinx' format.
}
\source{
Marcelo Araya-Salas
}
\usage{
data(selection_files)
}
\description{
List of 'Raven' and 'Syrinx' selection tables containing 
\emph{Phaethornis longirostris} (Long-billed Hermit) songs from the 
example sound files included in the \href{https://cran.r-project.org/package=warbleR}{warbleR} package (items 1 to 5),
a 'Raven' correlation matrix (item 6) and one selection from cane-brake wrens 
(item 7; no sound file available).
}
\keyword{datasets}
